pyproject.toml
[project]
name = "loukshop"
version = "0.1.0"
description = "DRF project for intership"
authors = [
    {name = "A.L",email = "abolfazlloukzade7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "djangorestframework-simplejwt (>=5.4.0,<6.0.0)",
    "django-autoslug (>=1.9.9,<2.0.0)",
    "django-filter (>=24.3,<25.0)",
    "pytz (>=2025.1,<2026.0)",
    "django-admin-interface (>=0.29.4,<0.30.0)",
    "pip (>=25.0.1,<26.0.0)",
    "drf-spectacular[sidecar] (>=0.28.0,<0.29.0)",
    "django-mptt (>=0.16.0,<0.17.0)",
    "django-ckeditor (>=6.7.2,<7.0.0)",
    "redis (>=5.2.1,<6.0.0)",
    "hiredis (>=3.1.0,<4.0.0)",
    "django-modeltranslation (>=0.19.12,<0.20.0)",
    "model-bakery (>=1.20.3,<2.0.0)",
    "drf-api-tracking (>=1.8.4,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
flake8 = "^7.1.1"

